% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{line_to_text}
\alias{line_to_text}
\title{Line to text}
\usage{
line_to_text(line = line)
}
\arguments{
\item{line}{the numeric value of the line: 6, 7, 8 or 9}
}
\value{

}
\description{
Converts a line value into a short text summary
}
\examples{
generate_hexagram() |>
line_to_text()

generate_hexagram() |>
dplyr::mutate(text = line_to_text(line))
}
