% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation.R
\name{line_to_text}
\alias{line_to_text}
\title{Line to text}
\usage{
line_to_text(line = line)
}
\arguments{
\item{line}{}
}
\value{

}
\description{
Converts a line value into a short text summary
}
\examples{
generate_hexagram() |>
line_to_text()

generate_hexagram() |>
mutate(text = line_to_text(line))
}
